WIP Prototype Mk10
Doing all the math shit was most likely unnecessary, as I simply had to remove the hardcoded chunk_scale in the chunk systems, but eh, having the scale tied to the identity of a chunk/chunk_loader/chunk_actor is probably still a smart idea. The math shit admittedly is a bit scuffed and redundant, but eh. Fuck it, it works, and I don't give a shit about anything else in the world right now.